+++ CHANGES.md
@@ -1,10 +1,12 @@
 ### dev
 
-[full changelog](http://github.com/yolk/valvat/compare/v0.0.2...0.0.3)
+[full changelog](http://github.com/yolk/valvat/compare/v0.0.3...master)
+
+* ActiveModel validation: added optional lookup support
 
 ### 0.0.3 / 2011-01-06
 
-[full changelog](http://github.com/yolk/valvat/compare/v0.0.1...v0.0.2)
+[full changelog](http://github.com/yolk/valvat/compare/v0.0.2...v0.0.3)
 
 * Basic support for ActiveModel validation